+++ packages/expo-av/android/src/main/java/expo/modules/av/AVManager.java
@@ -21,7 +21,6 @@ import org.unimodules.core.interfaces.InternalModule;
 import org.unimodules.core.interfaces.LifecycleEventListener;
 import org.unimodules.core.interfaces.services.EventEmitter;
 import org.unimodules.core.interfaces.services.UIManager;
-import org.unimodules.interfaces.permissions.Permissions;
 
 import java.io.File;
 import java.io.IOException;
@@ -38,6 +37,7 @@ import java.util.UUID;
 import expo.modules.av.player.PlayerData;
 import expo.modules.av.video.VideoView;
 import expo.modules.av.video.VideoViewWrapper;
+import expo.modules.interfaces.permissions.Permissions;
 
 import static android.media.MediaRecorder.MEDIA_RECORDER_INFO_MAX_FILESIZE_REACHED;