+++ packages/expo-facebook/ios/EXFacebook/EXFacebookAppTrackingPermissionRequester.h
@@ -1,7 +1,7 @@
 // Copyright 2017-present 650 Industries. All rights reserved.
 
-#import <UMPermissionsInterface/UMPermissionsInterface.h>
+#import <ExpoModulesCore/EXPermissionsInterface.h>
 
-@interface EXFacebookAppTrackingPermissionRequester : NSObject<UMPermissionsRequester>
+@interface EXFacebookAppTrackingPermissionRequester : NSObject<EXPermissionsRequester>
 
 @end