+++ packages/expo-image-picker/src/ExponentImagePicker.web.ts
@@ -1,5 +1,5 @@
 import { Platform } from '@unimodules/core';
-import { PermissionResponse, PermissionStatus } from 'unimodules-permissions-interface';
+import { PermissionResponse, PermissionStatus } from 'expo-modules-core';
 import { v4 } from 'uuid';
 
 import {