+++ packages/expo-location/ios/EXLocation/Requesters/EXBackgroundLocationPermissionRequester.m
@@ -65,30 +65,30 @@ static SEL alwaysAuthorizationSelector;
 
 - (NSDictionary *)parsePermissions:(CLAuthorizationStatus)systemStatus
 {
-  UMPermissionStatus status;
+  EXPermissionStatus status;
 
   switch (systemStatus) {
     case kCLAuthorizationStatusAuthorizedAlways: {
-      status = UMPermissionStatusGranted;
+      status = EXPermissionStatusGranted;
       break;
     }
     case kCLAuthorizationStatusDenied:
     case kCLAuthorizationStatusRestricted: {
-      status = UMPermissionStatusDenied;
+      status = EXPermissionStatusDenied;
       break;
     }
     case kCLAuthorizationStatusAuthorizedWhenInUse: {
       if (_wasAsked) {
-        status = UMPermissionStatusDenied;
+        status = EXPermissionStatusDenied;
       } else {
-        status = UMPermissionStatusUndetermined;
+        status = EXPermissionStatusUndetermined;
       }
       break;
     }
     case kCLAuthorizationStatusNotDetermined:
     default: {
       
-      status = UMPermissionStatusUndetermined;
+      status = EXPermissionStatusUndetermined;
       break;
     }
   }