+++ packages/expo-permissions/ios/EXPermissions/EXPermissions.m
@@ -1,12 +1,12 @@
 // Copyright 2016-present 650 Industries. All rights reserved.
 
-#import <UMPermissionsInterface/UMPermissionsInterface.h>
+#import <ExpoModulesCore/EXPermissionsInterface.h>
 
 #import <EXPermissions/EXPermissions.h>
 
 @interface EXPermissions ()
 
-@property (nonatomic, weak) id<UMPermissionsInterface> permissionsService;
+@property (nonatomic, weak) id<EXPermissionsInterface> permissionsService;
 
 @end
 
@@ -16,7 +16,7 @@ UM_EXPORT_MODULE(ExpoPermissions);
 
 - (void)setModuleRegistry:(UMModuleRegistry *)moduleRegistry
 {
-  _permissionsService = [moduleRegistry getModuleImplementingProtocol:@protocol(UMPermissionsInterface)];
+  _permissionsService = [moduleRegistry getModuleImplementingProtocol:@protocol(EXPermissionsInterface)];
 }
 
 # pragma mark - Exported methods
@@ -26,7 +26,7 @@ UM_EXPORT_METHOD_AS(getAsync,
                     resolver:(UMPromiseResolveBlock)resolve
                     rejecter:(UMPromiseRejectBlock)reject)
 {
-  id<UMPermissionsRequester> requester = [_permissionsService getPermissionRequesterForType:permissionType];
+  id<EXPermissionsRequester> requester = [_permissionsService getPermissionRequesterForType:permissionType];
   if (requester == nil) {
     return reject(@"E_PERMISSIONS_UNKNOWN", [NSString stringWithFormat:@"Unrecognized permission: %@", permissionType], nil);
   }
@@ -40,7 +40,7 @@ UM_EXPORT_METHOD_AS(askAsync,
                     resolver:(UMPromiseResolveBlock)resolve
                     rejecter:(UMPromiseRejectBlock)reject)
 {
-  id<UMPermissionsRequester> requester = [_permissionsService getPermissionRequesterForType:permissionType];
+  id<EXPermissionsRequester> requester = [_permissionsService getPermissionRequesterForType:permissionType];
   if (requester == nil) {
     return reject(@"E_PERMISSIONS_UNKNOWN", [NSString stringWithFormat:@"Unrecognized permission: %@", permissionType], nil);
   }