+++ packages/expo-ads-admob/CHANGELOG.md
@@ -14,6 +14,7 @@
 ### 💡 Others
 
 - Build Android code using Java 8 to fix Android instrumented test build error. ([#12939](https://github.com/expo/expo/pull/12939) by [@kudo](https://github.com/kudo))
+- Migrated from `unimodules-permissions-interface` to `expo-modules-core`. ([#12961](https://github.com/expo/expo/pull/12961) by [@tsapeta](https://github.com/tsapeta))
 
 ## 10.0.4 — 2021-04-13