+++ packages/expo-barcode-scanner/CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ### 💡 Others
 
-- Migrated module interface from `unimodules-barcode-scanner-interface` to `expo-modules-core`. ([#12912](https://github.com/expo/expo/pull/12912) by [@tsapeta](https://github.com/tsapeta))
+- Migrated interfaces from their own packages to `expo-modules-core`. ([#12912](https://github.com/expo/expo/pull/12912) by [@tsapeta](https://github.com/tsapeta))
 
 ## 10.1.2 — 2021-04-13